refactor: redistribute cmd deploy functions and files
